Prune redundant public client tests (#608)
